Rename IterTools to IteratorTools to avoid problems with mac
